+++ src/epc/FileCoreProperties.cpp
@@ -91,7 +91,7 @@ std::string FileCoreProperties::toString() const
 	oss << "xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\">" << endl;
 
 	// content
-	for (unsigned int i = 0; i < CoreProperty::undefinedCoreProperty; ++i) {
+	for (size_t i = 0; i < CoreProperty::undefinedCoreProperty; ++i) {
 		if (!properties[i].isEmpty()) {
 			oss << "\t" << properties[i].toString() << endl;
 		}
@@ -185,7 +185,7 @@ void FileCoreProperties::setVersion(const std::string & value)
 
 void FileCoreProperties::readFromString(const string & textInput)
 {
-	for (unsigned int i = 0; i < CoreProperty::undefinedCoreProperty; ++i) {
+	for (size_t i = 0; i < CoreProperty::undefinedCoreProperty; ++i) {
 		properties[i].setTypeProperty(CoreProperty::undefinedCoreProperty);
 	}
